formated

+++ app/api/src/module/user/user.service.ts
@@ -68,18 +68,12 @@ export class UserService {
     return commonFriends;
   }
 
-  async DeleteFriend(
-    friends: string,
-    name: string,
-  ): Promise<UserGetDto> {
-    if (friends)
-      return this.repository.deleteFriend(name, friends);
+  async DeleteFriend(friends: string, name: string): Promise<UserGetDto> {
+    if (friends) return this.repository.deleteFriend(name, friends);
   }
 
-  async DeleteUser(
-    name: string,
-  ): Promise<UserGetDto> {
-      return this.repository.deleteUser(name);
+  async DeleteUser(name: string): Promise<UserGetDto> {
+    return this.repository.deleteUser(name);
   }
 
   async CheckFriendsUpdate(friend: string, name: string): Promise<void> {
